+++ pom.xml
@@ -6,7 +6,7 @@
 
 	<groupId>ro.brage</groupId>
 	<artifactId>dodo</artifactId>
-	<version>2.0.0</version>
+	<version>3.0.0</version>
 	<packaging>jar</packaging>
 	<name>dodo</name>
